Move Retest::Program to its own file
